+++ sdk/storage/Azure.Storage.DataMovement/src/StorageResource.cs
@@ -18,16 +18,6 @@ namespace Azure.Storage.DataMovement
         }
 
         /// <summary>
-        /// Defines whether the storage resource type can produce a web URL.
-        /// </summary>
-        protected internal abstract bool CanProduceUri { get; }
-
-        /// <summary>
-        /// Gets the Uri.
-        /// </summary>
-        public abstract Uri Uri { get; }
-
-        /// <summary>
         /// Gets the path.
         /// </summary>
         public abstract string Path { get; }
@@ -36,5 +26,16 @@ namespace Azure.Storage.DataMovement
         /// Defines whether the storage resource is a container.
         /// </summary>
         protected internal abstract bool IsContainer { get; }
+
+        /// <summary>
+        /// Attempts to get the Uri of the Storage Resource.
+        /// </summary>
+        /// <param name="uri">
+        /// The <see cref="Uri"/> of the Storage Resource if the Storage Resource has one.
+        /// </param>
+        /// <returns>
+        /// Returns true if retrieving the <see cref="Uri"/> was successful. Returns false if not.
+        /// </returns>
+        public abstract bool TryGetUri(out Uri uri);
     }
 }