+++ sdk/storage/Azure.Storage.DataMovement/tests/BlockBlobStorageResourceTests.cs
@@ -32,9 +32,12 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             BlockBlobStorageResource storageResource = new BlockBlobStorageResource(blobClient);
 
             // Assert
-            Assert.AreEqual(uri, storageResource.Uri);
+            if (!storageResource.TryGetUri(out Uri storageUri))
+            {
+                throw Errors.ResourceUriInvalid(nameof(storageResource));
+            }
+            Assert.AreEqual(uri, storageUri.AbsoluteUri);
             Assert.AreEqual(blobClient.Name, storageResource.Path);
-            Assert.IsTrue(storageResource.CanProduceUri);
         }
 
         [RecordedTest]