+++ sdk/storage/Azure.Storage.DataMovement/tests/PageBlobStorageResourceTests.cs
@@ -33,9 +33,12 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             PageBlobStorageResource storageResource = new PageBlobStorageResource(blobClient);
 
             // Assert
-            Assert.AreEqual(uri, storageResource.Uri);
+            if (!storageResource.TryGetUri(out Uri storageUri))
+            {
+                throw Errors.ResourceUriInvalid(nameof(storageResource));
+            }
+            Assert.AreEqual(uri, storageUri.AbsoluteUri);
             Assert.AreEqual(blobClient.Name, storageResource.Path);
-            Assert.IsTrue(storageResource.CanProduceUri);
         }
 
         [RecordedTest]